+++ Packages/com.ishix.nativeprompt/Scripts/Runtime/IOSNativePromptStrategy.cs
@@ -86,6 +86,11 @@ namespace NativePrompt
                 options.BackgroundColor.g,
                 options.BackgroundColor.b,
                 options.BackgroundOpacity,
+                options.MessageColor.r,
+                options.MessageColor.g,
+                options.MessageColor.b,
+                options.MessageColor.a,
+                options.MessageFontSize,
                 (int)options.Position,
                 (int)options.Size,
                 options.ShowDelaySeconds);
@@ -211,6 +216,11 @@ namespace NativePrompt
             float backgroundGreen,
             float backgroundBlue,
             float backgroundOpacity,
+            float messageRed,
+            float messageGreen,
+            float messageBlue,
+            float messageAlpha,
+            float messageFontSize,
             int position,
             int size,
             float showDelaySeconds);